Revert FC React renaming
