+++ app/layouts/__test__/default-layout.test.tsx
@@ -3,20 +3,18 @@
 import { render } from "@testing-library/react";
 import { MemoryRouter } from "react-router";
 import { it } from "vitest";
-import NarrowLayout from "../NarrowLayout";
+import DefaultLayout from "../default-layout";
 
-describe("NarrowLayout", () => {
+describe("DefaultLayout", () => {
   it("should render the expected layout with <main/> and <footer/> DOM elements", () => {
     const { getByRole, getByTestId } = render(
       <MemoryRouter>
-        <NarrowLayout />
+        <DefaultLayout />
       </MemoryRouter>,
     );
 
     // check if correct class is present
-    expect(getByTestId("narrow-layout")).toHaveClass(
-      "kern-col-sm-10 kern-col-md-8 kern-col-lg-6 kern-col-xxl-4",
-    );
+    expect(getByTestId("default-layout")).toHaveClass("kern-container");
     // check if main is rendered
     expect(getByRole("main")).toBeInTheDocument();
     // check if footer is rendered
